+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.7.0]
+
+### Added
+- Introduced `exh.notificationsV2` with methods for Notifications Service v2:
+  - `exh.notificationsV2.*` provides functionality to send and retrieve notifications
+  - `exh.notificationsV2.userSettings.*` offers management of user's devices
+
 ## [8.6.0]
 
 ### Added